requirements.txt
click
gitpython
requests
pygithub3
